requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
python-dotenv==1.0.1